refactor: 결제 금액 자료형 BigDecimal로 변경

+++ src/main/java/roomescape/dto/response/theme/ThemePriceResponse.java
@@ -1,4 +1,6 @@
 package roomescape.dto.response.theme;
 
-public record ThemePriceResponse(int price) {
+import java.math.BigDecimal;
+
+public record ThemePriceResponse(BigDecimal price) {
 }
